PostsPage created and app.js has been redone to reflect change
