+++ instagram/src/App.js
@@ -1,46 +1,17 @@
 import React, { Component } from 'react';
 import './App.css';
-import dummyData from './dummy-data';
-import SearchBar from './Components/SearchBar/SearchBar';
-import PostsContainer from './Components/PostContainer/PostContainer';
+import PostsPage from './Components/PostContainer/PostsPage';
 
 class App extends Component {
   constructor() {
     super();
-    this.state = {
-      posts: [],
-      filteredPosts: []
-    };
+    this.state = {};
   }
-
-  componentDidMount() {
-    this.setState({ posts: dummyData });
-  }
-  searchPostsHandler = e => {
-    const posts = this.stats.posts.filter(p => {
-      if (p.username.includes(e.target.value)) {
-        return p;
-      }
-    });
-    this.setState({ filteredPosts: posts });
-  };
-
-
 
   render() {
     return (
       <div className="App">
-        <SearchBar 
-          searchTerm={this.state.searchTerm}
-          searchPosts={this.searchPostsHandler}
-        />
-        <PostsContainer 
-          posts={
-            this.state.filteredPosts.length > 0
-            ? this.state.filteredPosts 
-            : this.state.posts
-          }
-        />
+        <PostsPage />
       </div>
     );
   }
