+++ instagram/src/Components/PostContainer/PostsPage.js
@@ -1,0 +1,49 @@
+import React, { Component } from 'react';
+
+import dummyData from '../../dummy-data';
+import SearchBar from '../SearchBar/SearchBar';
+import PostsContainer from './PostContainer';
+
+class PostsPage extends Component {
+    constructor() {
+      super();
+      this.state = {
+        posts: [],
+        filteredPosts: []
+      };
+    }
+  
+    componentDidMount() {
+      this.setState({ posts: dummyData });
+    }
+    searchPostsHandler = e => {
+      const posts = this.stats.posts.filter(p => {
+        if (p.username.includes(e.target.value)) {
+          return p;
+        }
+      });
+      this.setState({ filteredPosts: posts });
+    };
+  
+  
+  
+    render() {
+      return (
+        <div className="App">
+          <SearchBar 
+            searchTerm={this.state.searchTerm}
+            searchPosts={this.searchPostsHandler}
+          />
+          <PostsContainer 
+            posts={
+              this.state.filteredPosts.length > 0
+              ? this.state.filteredPosts 
+              : this.state.posts
+            }
+          />
+        </div>
+      );
+    }
+  }
+
+  export default PostsPage;
